requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.4.0
anthropic==0.40.0
httpx==0.27.2
pyyaml==6.0.2
python-dotenv==1.0.1
